pyserver/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
mediapipe==0.10.18
numpy==2.1.3
opencv-python==4.10.0.84
Pillow==11.0.0
